backend/hacknu24/requirements.txt
Django==4.2.7
django-cors-headers==4.3.0
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
phonenumbers==8.13.34